accout 앱 제작, account 적용

+++ Blog/BlogApp/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Blog
 from .form import PostForm
 from django.utils import timezone
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def post_list(request):
@@ -13,41 +14,35 @@ def detail(request, post_id):
     post = get_object_or_404(Blog, pk =  post_id)
     return render(request, 'detail.html', {'post':post})
 
-
+@login_required(login_url='/account/login/')
 def new(request):
     if request.method == 'POST':
         form = PostForm(request.POST, request.FILES)
         if form.is_valid():
             content = form.save(commit = False)
             content.published_at = timezone.datetime.now()
+            content.publisher = request.user.username
             content.save()
             return redirect('post_list')
     else:
         form = PostForm()
         return render(request, 'new.html', {'form':form})
 
-
-def create(request):
-    new_blog = Blog()
-    new_blog.title = request.POST['title']
-    new_blog.body = request.POST['body']
-    new_blog.publisher = request.POST['publisher']
-    new_blog.rep_img = request.FILES['rep_img']
-    new_blog.hashtag = request.POST['hashtag']
-    new_blog.published_at = timezone.datetime.now()
-    new_blog.save()
-    return redirect('/blog/'+str(new_blog.id))
-
-
+@login_required(login_url='/account/login/')
 def edit(request, post_id):
+    user = request.user
     form = PostForm(instance=Blog.objects.get(pk=post_id))
-    return render(request, 'edit.html', {'form':form, 'post_id':post_id})
+    if user.is_authenticated and Blog.objects.get(pk=post_id).publisher == user.username:
+        return render(request, 'edit.html', {'form':form, 'post_id':post_id})
+    elif user.is_authenticated:
+        return redirect('/')
+    else:
+        return redirect("login")
 
 def update(request, post_id):
     update_blog = get_object_or_404(Blog, pk = post_id)
     update_blog.title = request.POST['title']
     update_blog.body = request.POST['body']
-    update_blog.publisher = request.POST['publisher']
     try:
         update_blog.rep_img = request.FILES['rep_img']
     except:
@@ -56,8 +51,14 @@ def update(request, post_id):
     update_blog.save()
     return redirect('detail', update_blog.id)
 
-
+@login_required(login_url='/account/login/')
 def delete(request,post_id):
-    delete_blog  = get_object_or_404(Blog, pk = post_id)
-    delete_blog.delete()
-    return redirect('/')
+    user = request.user
+    if user.is_authenticated and Blog.objects.get(pk=post_id).publisher == user.username:
+        delete_blog  = get_object_or_404(Blog, pk = post_id)
+        delete_blog.delete()
+        return redirect('/')
+    elif user.is_authenticated:
+        return redirect('/')
+    else:
+        return redirect("login")
